requirements.txt
click>=8.2.1
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
werkzeug>=3.1.3
firebase-admin>=6.2.0
pyrebase4>=4.7.1
python-dotenv>=1.0.0
